+++ public/less/common.less
@@ -991,3 +991,9 @@ wz-xml-file-editor {
 discover-app-w .container-fluid {
   background: #fff;
 }
+
+.registerAgent{
+  min-height: calc(~'100vh - 100px');
+  background: #fafbfd;
+  padding-top: 25px;
+}
